requirements.txt
fastapi
uvicorn
sqlalchemy
requests
python-dotenv
pandas
google-genai
